Adding operator and function to convert to set layout binding
And fix a couple warnings and minor things, like storing binding index as a size_t (when its a 32-bit uint)

+++ src/parser/yamlFile.cpp
@@ -298,7 +298,7 @@ namespace st {
                     rsrc.SetQualifiers(qualifiers.size(), qualifiers.data());
                 }
 
-                rsrc.SetBindingIndex(group_resources.size());
+                rsrc.SetBindingIndex(static_cast<uint32_t>(group_resources.size()));
                 group_resources.emplace_back(std::move(rsrc));
 
             }
